url_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
psycopg2-binary==2.9.10
redis==5.2.1
confluent-kafka==2.3.0
pydantic>=2.0
